Translations, typos, and HTML is now marked safe
Form labels with HTML are now hard-coded in the form Meta classes, rather than being pulled from the model's field description. This is necessary because of an interaction with SoftValidationMixin which for some reason escapes all labels (needs research/better fix). Notes have been added in model description to prevent confusion in the future.

TODO: Some fixtures and their hard-coded translations were modified, meaning these might need to be modified in-place as well.