Fix/pre proposals permission denied (571) (#577)
* fix: Remove custom form_valid from ProposalSubmitPreApproved

The parent's form_valid was already being run by fetching it's output
as a success_url. Then a copy of that code was run locally, which
would raise a PermissionDenied because the proposal was already
submitted.

Because this version did nothing different, except cause an error, it
was removed.

* fix: Do not call parent form_valid in ProposalSubmitPreAssessment

The previously existing comment pointed out that this should not be
done, then the method did it anyway. The call to the parent's
form_valid was removed and the comment tweaked.

* fix: return a form_valid response rather than just the success_url

* feature: Break out the submission actions into their own method

* feat: consistently return a review in all start_review functions

* feat: Handle practice, supervised, and preassessment proposals in start_review

* feat: remove proposal-specific code in form_valid and explain why

* fix: slightly better comment

* feat: Move BaseViewTestCase to main app

* feat: Tests for proposal submission

Proposal tests were also moved to their own directory

* fix: BaseViewTestCase no longer explicitly references reviews

* fix: Fetch csrf token for any user, not just secretary

* feat: Testing submission of regular proposals

* feat: Test submission of the two alternate proposal types

* fix: Check submission status codes, and add comments

* fix: Rename the other BaseProposalTestCase to MiscProposalTestCase

* fix: Supply needed user for logged-in post() method

* fix: Typo in comment

* Load test proposals through fixtures system

* feat: New command to load all _initial_ fixtures

Moved testing fixtures to testing subdir